requirements.txt
Django>=1.8,<2.0
djangorestframework==3.8.2
pytz==2018.4
mysqlclient==1.3.12
